Update structure of controllers
Update structure of controllers for contacts.
