+++ controllers/contacts/deleteContact.js
@@ -1,0 +1,17 @@
+const { Contact } = require("../../models/contact");
+const { HttpError, ctrlWrapper } = require("../../helpers");
+
+const deleteContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndRemove(contactId);
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+  res.json({
+    message: "contact deleted",
+  });
+};
+
+module.exports = {
+  deleteContact: ctrlWrapper(deleteContact),
+};
