+++ controllers/contacts/getAllContacts.js
@@ -1,0 +1,10 @@
+const { Contact } = require("../../models/contact");
+const { ctrlWrapper } = require("../../helpers");
+
+const getAllContacts = async (req, res) => {
+  res.json(await Contact.find());
+};
+
+module.exports = {
+  getAllContacts: ctrlWrapper(getAllContacts),
+};
