--- controllers/controllers.js
@@ -1,58 +1,0 @@
-const { Contact } = require("../models/contact");
-const { HttpError, ctrlWrapper } = require("../helpers");
-
-const getAllContacts = async (req, res) => {
-  res.json(await Contact.find());
-};
-
-const getById = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(result);
-};
-
-const addNewContact = async (req, res) => {
-  const result = await Contact.create(req.body);
-  res.status(201).json(result);
-};
-
-const deleteContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json({
-    message: "contact deleted",
-  });
-};
-
-const changeContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {new: true});
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(result);
-};
-
-const changeFavorite = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {new: true});
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(result);
-}
-
-module.exports = {
-  getAllContacts: ctrlWrapper(getAllContacts),
-  getById: ctrlWrapper(getById),
-  addNewContact: ctrlWrapper(addNewContact),
-  deleteContact: ctrlWrapper(deleteContact),
-  changeContact: ctrlWrapper(changeContact),
-  changeFavorite: ctrlWrapper(changeFavorite),
-};
