+++ routes/api/contacts.js
@@ -1,20 +1,28 @@
 const express = require('express');
 const router = express.Router();
 
-const controllers = require('../../controllers/controllers');
+const {
+    changeContact,
+    deleteContact,
+    addNewContact,
+    getById,
+    getAllContacts,
+    changeFavorite
+} = require('../../controllers/contacts');
+
 const { validateBody, isValidId } = require('../../middlewares');
 const { schemas } = require("../../models/contact");
 
-router.get('/', controllers.getAllContacts);
+router.get('/', getAllContacts);
 
-router.get('/:contactId', isValidId, controllers.getById);
+router.get('/:contactId', isValidId, getById);
 
-router.post('/', validateBody(schemas.addSchema), controllers.addNewContact);
+router.post('/', validateBody(schemas.addSchema), addNewContact);
 
-router.delete('/:contactId', isValidId, controllers.deleteContact);
+router.delete('/:contactId', isValidId, deleteContact);
 
-router.put('/:contactId', isValidId, validateBody(schemas.addSchema), controllers.changeContact);
+router.put('/:contactId', isValidId, validateBody(schemas.addSchema), changeContact);
 
-router.patch('/:contactId/favorite', isValidId, validateBody(schemas.changeFavoriteSchema), controllers.changeFavorite);
+router.patch('/:contactId/favorite', isValidId, validateBody(schemas.changeFavoriteSchema), changeFavorite);
 
 module.exports = router;
